refactor: create routes and improve documentation
