+++ app/database/connection.py
@@ -1,6 +1,8 @@
 import os
+
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, declarative_base
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
 
 mysql_username = os.getenv('MYSQL_USERNAME')
 mysql_password = os.getenv('MYSQL_PASSWORD')
@@ -15,3 +17,10 @@ engine = create_engine(
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
+
+def get_database():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
